requirements.txt
PyMuPDF==1.23.26
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
requests==2.31.0
sentence-transformers==2.5.1
tqdm==4.66.2
transformers==4.38.2
faiss-cpu
streamlit
torch
setuptools
wheel
